# docs/swagger.yaml
basePath: /api/v1
definitions:
  httpErrors.RestErr:
    type: object
...
  models.Comment:
    properties:
      author_id:
        type: string
      comment_id:
        type: string
      created_at:
        type: string
      likes:
        type: integer
      message:
        type: string
      news_id:
        type: string
      updated_at:
        type: string
    required:
    - author_id
    - message
    - news_id
    type: object
  models.CommentBase:
    properties:
      author:
        type: string
      author_id:
        type: string
      avatar_url:
        type: string
      comment_id:
        type: string
      likes:
        type: integer
      message:
        type: string
      updated_at:
        type: string
    required:
    - author
    - author_id
    - message
    type: object
  models.CommentsList:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.CommentBase'
        type: array
      has_more:
        type: boolean
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.News:
    properties:
      author_id:
        type: string
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      image_url:
        type: string
      news_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - author_id
    - content
    - title
    type: object
  models.NewsList:
    properties:
      has_more:
        type: boolean
      news:
        items:
          $ref: '#/definitions/models.News'
        type: array
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact:
    email: alexander.bryksin@yandex.ru
    name: Alexander Bryksin
    url: https://github.com/AleksK1NG
  description: Example Golang REST API
  title: Go Example REST API
  version: "1.0"
paths:
...
  /comments:
    post:
      consumes:
      - application/json
      description: create new comment
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      summary: Create new comment
      tags:
      - Comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: delete comment
      parameters:
      - description: comment_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      summary: Delete comment
      tags:
      - Comments
    get:
      consumes:
      - application/json
      description: Get comment by id
      parameters:
      - description: comment_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      summary: Get comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: update new comment
      parameters:
      - description: comment_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      summary: Update comment
      tags:
      - Comments
  /comments/byNewsId/{id}:
    get:
      consumes:
      - application/json
      description: Get all comment by news id
      parameters:
      - description: news_id
        in: path
        name: id
        required: true
        type: integer
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommentsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      summary: Get comments by news
      tags:
      - Comments
  /news:
    get:
      consumes:
      - application/json
      description: Get all news with pagination
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NewsList'
      summary: Get all news
      tags:
      - News
  /news/{id}:
    delete:
      consumes:
      - application/json
      description: Delete by id news handler
      parameters:
      - description: news_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Delete news
      tags:
      - News
    get:
      consumes:
      - application/json
      description: Get by id news handler
      parameters:
      - description: news_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: Get by id news
      tags:
      - News
    put:
      consumes:
      - application/json
      description: Update news handler
      parameters:
      - description: news_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: Update news
      tags:
      - News
...
  /news/search:
    get:
      consumes:
      - application/json
      description: Search news by title
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NewsList'
      summary: Search by title
      tags:
      - News
swagger: "2.0"